Refactor: Makefile, Core 리팩토링
- make re 수행시, 불필요한 과정 제거.
- Core::handleLogBuffer()에서, if((uint64)wrote == len) 으로 코드 변경.
  - wrote = uint64로 되어 있어서, 서로 다른 type간 비교로 컴파일 오류
  발생.
  - 비교시에만 int64->uint64로 형변환 하도록 수정.

+++ src/grc/core/Core.cpp
@@ -259,7 +259,7 @@ void Core::handleLogBuffer()
     }
 
     mLogBufferIndex += wrote;
-    if (wrote == len)
+    if ((uint64)wrote == len)
     {
         mLogBuffer.clear();
         mLogBufferIndex = 0;
